move spill related into new file
